+++ practice 4/src/main/resources/db/V2022.10.31.21.01__create.sql
@@ -7,8 +7,8 @@ CONSTRAINT roles_pk PRIMARY KEY (id)
 CREATE TABLE permissions (
 id SERIAL NOT NULL,
 name varchar(64) NOT NULL,
-role_id bigint REFERENCES roles(id),
+role_id SERIAL REFERENCES roles(id),
 CONSTRAINT permissions_pk PRIMARY KEY (id)
 );
 
-ALTER TABLE users ADD COLUMN role_id bigint REFERENCES roles(id);
+ALTER TABLE users ADD COLUMN role_id SERIAL REFERENCES roles(id);
